Use GuidProvider and DateTimeProvider
This improves the testability of code that involves Chorus.
