+++ src/LibChorus/notes/Annotation.cs
@@ -11,6 +11,7 @@ using System.Xml.XPath;
 using Chorus.Properties;
 using Chorus.Utilities;
 using Chorus.merge.xml.generic;
+using Palaso.Providers;
 
 namespace Chorus.notes
 {
@@ -29,7 +30,7 @@ namespace Chorus.notes
         }
 
 		public Annotation(string annotationClass, string refUrl, string path)
-			: this(annotationClass, refUrl, System.Guid.NewGuid(), path)
+			: this(annotationClass, refUrl, GuidProvider.Current.NewGuid(), path)
 		{
 		}
